requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-cors-headers==3.0.2
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-mysql==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.8.2
env==0.1.0
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.2
Pillow==5.3.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
stripe==2.32.1
urllib3==1.25.3
wincertstore==0.2
